+++ Sources/RichTextKit/Attributes/RichTextAttributeReader.swift
@@ -52,7 +52,7 @@ public extension RichTextAttributeReader {
         at range: NSRange
     ) -> RichTextAttributes {
         if richText.length == 0 { return [:] }
-        let range = safeRange(for: range)
+        let range = safeRange(for: range, isAttributeOperation: true)
         return richText.attributes(at: range.location, effectiveRange: nil)
     }
 }