+++ Sources/RichTextKit/Indent/RichTextIndentWriter.swift
@@ -106,7 +106,7 @@ private extension RichTextIndentWriter {
     ) -> RichTextAttributes? {
         guard let text = mutableRichText else { return nil }
         let range = NSRange(location: index, length: 1)
-        let safeRange = safeRange(for: range)
+        let safeRange = safeRange(for: range, isAttributeOperation: true)
         var attributes = text.attributes(at: safeRange.location, effectiveRange: nil)
         let style = attributes[.paragraphStyle] as? NSMutableParagraphStyle ?? NSMutableParagraphStyle()